Add fabric8io/docker-maven-plugin for creating the docker images
The Docker build configuration creates the same Docker image as declared in Dockerfile. In addition it already adds the proper name for the image (not possible to specify in the Dockerfile) + it would be able to push it to a registry.
Also it automatically updates to version upgrades, where a plain Dockerfile needs to be updared manually (there is an error btw in the current Dockerfile, it refers to the non-fat jar file in its ADD).

A sample run configuration is added which maps 8080 to a dynamic port and then assigns it to a Maven property ${aloha.port} which could be used in integration tests (along with ${docker.host.address} which is set to the IP of the Docker daemon).
The <run> section is optional and can be removed if not needed.